requirements.txt
pyyaml>=6.0
google-generativeai>=0.8.0
pillow>=10.0.0
